Fleshed out cabal configuration with a skeleton!
`cabal init` tries hard to get the project into a good initial state but
needs a little further help to get entirely up to speed. Package
configuration will obviously continue to develop as the project grows, but
at this point we want to be able to exercise most of the cabal commands
successfully. We could even upload this to Hackage now, though it'd be sort
of annoying.