Added Haddock comments to **/Parser.hs
Haddock is Haskell's (quite nice) integrated documentation system. It's used
to automatically generate documentation for all of Hackage and it's often
much nicer to read than plain code---better formatting, hyperlinks, the
whole nine yards.

Documentation is fascinating in Haskell since so much information can be
placed into the type signature. Often I try to only document the intent of a
function and let the type signature show you how to use it. More complex
signatures might deserve some doctests or specializations, but that probably
won't occur in as simple a library as this.