Created a first-draft implementation of Parser
Right now `Parser` defers defining its Monad using
GeneralizedNewtypeDeriving to pick up the instance from IO. Alas, since
we've only layered `IO` into Parser so far it has no way to fail except to
throw a runtime exception... That isn't the *worst* that could happen since
if you're looking for something in the environment and it just isn't
there... you're probably sunk for running the rest of your application and
might as well call it a day.

But the whole point of this library is that there's something better you can
do.