Added a QuickCheck-style test as well
QuickCheck style tests are property tests--tests which are run over an
unknown set of test data which is randomly generated on the fly. This
creates a very different style of testing where the test writer must thing
about properties which hold over arbitrary data. This can be a *very*
powerful way of thinking about your code and it's one that Haskell supports
quite a lot.

For instance, the test added here checks that, for all names, if we store a
value in the ENV at that name and then get it back out using our Env.Parser
that the value is intact. To do this the property test randomly generates
100 names and values and perhaps the test for each one.

I don't think that's actually a super high value test---normally, I'd be
more interested in something like this when I suspect there might be edge
cases buried in the sea of random data. In this case, it serves more as a
placeholder and example.