Env.doc: static, pure Parser documentation
Haskell being, lazy and declarative, can easily expand the meaning of a type
by adjoining the alternative semantics and piping it through all the same
functions as before. This is a little bit heavy-weight and won't scale, but
it's very easy to do.

Here we add a completely pure semantics to Env.Parser which computes the
environment variables the parser depends upon.