Free Applicative impl of Env.Parser
Now that we have our free applicative library embedded, we can use it
to easily reimplement the multiple behaviors of the Env.Parser. The
advantage of this method is that it's much more explicit and easier to
extend both in providing new interpreters and extending the signature
of the Parser should we wish to add more effects.