Slots now carry around ways to show their defaults
Previously it'd be difficult to summarize a tree and carry along defaults.
Each default value could be of a different type and thus we couldn't make a
homogenous list and if we made a heterogenous list then we'd lose the
ability to show the defaults.

To remedy this the Slot interface has been changed so that it always has a
chance to provide a String representation of the default value---either
through that value's Show instance or by explicitly providing it using a
lens.